+++ cookbooks/chef-ingredient/libraries/chef_ingredient_provider.rb
@@ -16,13 +16,12 @@
 #
 
 require_relative './helpers'
-require_relative './debian_handler'
-require_relative './rhel_handler'
+require_relative './default_handler'
 require_relative './omnitruck_handler'
 
 class Chef
   class Provider
-    class ChefIngredient < Chef::Provider::LWRPBase
+    class ChefIngredient < Chef::Provider::LWRPBase # ~FC058
       provides :chef_ingredient
       use_inline_resources
 
@@ -38,16 +37,17 @@ class Chef
       def initialize(name, run_context = nil)
         super(name, run_context)
         case node['platform_family']
-        when 'debian'
-          extend ::ChefIngredient::DebianHandler
-        when 'rhel'
-          extend ::ChefIngredient::RhelHandler
+        when 'debian', 'rhel', 'windows'
+          extend ::ChefIngredient::DefaultHandler
         else
+          # OmnitruckHandler is used for Solaris, AIX, FreeBSD, etc.
+          # Eventually, we would like to support all platforms with the DefaultHandler
           extend ::ChefIngredient::OmnitruckHandler
         end
       end
 
       action :install do
+        check_deprecated_properties
         add_config(new_resource.product_name, new_resource.config)
         declare_chef_run_stop_resource
 
@@ -55,6 +55,7 @@ class Chef
       end
 
       action :upgrade do
+        check_deprecated_properties
         add_config(new_resource.product_name, new_resource.config)
         declare_chef_run_stop_resource
 
@@ -62,10 +63,12 @@ class Chef
       end
 
       action :uninstall do
+        check_deprecated_properties
         handle_uninstall
       end
 
       action :reconfigure do
+        check_deprecated_properties
         add_config(new_resource.product_name, new_resource.config)
 
         if ingredient_ctl_command.nil?
@@ -76,6 +79,27 @@ class Chef
           ingredient_config new_resource.product_name do
             action :render
             not_if { get_config(new_resource.product_name).empty? }
+          end
+
+          # If accept_license is set, drop .license.accepted file so that
+          # reconfigure does not prompt for license acceptance. This is
+          # the backwards compatible way of accepting a Chef license.
+          if new_resource.accept_license && %w(analytics manage reporting compliance).include?(new_resource.product_name)
+            # The way we construct the data directory for a product, that looks
+            # like /var/opt/<product_name> is to get the config file path that
+            # looks like /etc/<product_name>/<product_name>.rb and do path
+            # manipulation.
+            product_data_dir_name = ::File.basename(::File.dirname(ingredient_config_file(new_resource.product_name)))
+            product_data_dir = ::File.join('/var/opt', product_data_dir_name)
+
+            directory product_data_dir do
+              recursive true
+              action :create
+            end
+
+            file ::File.join(product_data_dir, '.license.accepted') do
+              action :touch
+            end
           end
 
           execute "#{ingredient_package_name}-reconfigure" do