+++ cookbooks/chef-ingredient/libraries/omnitruck_handler.rb
@@ -27,8 +27,8 @@ module ChefIngredient
         else
           Chef::Log.debug("Found version #{current_version}, skipping installing :latest.")
         end
-      else
-        configure_version(installer) if new_resource.version != current_version
+      elsif new_resource.version != current_version
+        configure_version(installer)
       end
     end
 
@@ -37,57 +37,44 @@ module ChefIngredient
     end
 
     def handle_uninstall
-      fail 'Uninstalling a product is currently not supported.'
+      raise 'Uninstalling a product is currently not supported.'
     end
 
     def configure_version(installer)
       install_command_resource = "install-#{new_resource.product_name}-#{new_resource.version}"
 
-      file installer_script_path do
-        content installer.install_command
-        if windows?
+      if windows?
+        file installer_script_path do
+          content installer.install_command
           notifies :run, "powershell_script[#{install_command_resource}]", :immediately
-        else
+        end
+
+        powershell_script install_command_resource do
+          # We pass the install code directly, but still depend upon the file to
+          # change before executing the install
+          code installer.install_command
+          action :nothing
+
+          if new_resource.product_name == 'chef'
+            # We define this resource in ChefIngredientProvider
+            notifies :run, 'ruby_block[stop chef run]', :immediately
+          end
+        end
+      else
+        file installer_script_path do
+          content installer.install_command
           notifies :run, "execute[#{install_command_resource}]", :immediately
         end
-      end
 
-      powershell_script install_command_resource do
-        # We pass the install code directly, but still depend upon the file to
-        # change before executing the install
-        code installer.install_command
-        action :nothing
+        execute install_command_resource do
+          command "sudo /bin/sh #{installer_script_path}"
+          action :nothing
 
-        if new_resource.product_name == 'chef'
-          # We define this resource in ChefIngredientProvider
-          notifies :run, 'ruby_block[stop chef run]', :immediately
+          if new_resource.product_name == 'chef'
+            # We define this resource in ChefIngredientProvider
+            notifies :run, 'ruby_block[stop chef run]', :immediately
+          end
         end
-      end
-
-      execute install_command_resource do
-        command "sudo /bin/sh #{installer_script_path}"
-        action :nothing
-
-        if new_resource.product_name == 'chef'
-          # We define this resource in ChefIngredientProvider
-          notifies :run, 'ruby_block[stop chef run]', :immediately
-        end
-      end
-    end
-
-    def installer
-      @installer ||= begin
-        ensure_mixlib_install_gem_installed!
-
-        options = {
-          product_name: new_resource.product_name,
-          channel: new_resource.channel,
-          product_version: new_resource.version
-        }.tap do |opt|
-          opt[:shell_type] = :ps1 if windows?
-        end
-
-        Mixlib::Install.new(options)
       end
     end