+++ cookbooks/chef-ingredient/test/fixtures/cookbooks/test/metadata.rb
@@ -1,0 +1,8 @@
+name 'test'
+version '0.0.1'
+
+depends 'chef-ingredient'
+depends 'custom_repo'
+depends 'git'
+depends 'apt'
+depends 'yum'