+++ cookbooks/chef-ingredient/test/fixtures/cookbooks/test/recipes/unstable_chef_server.rb
@@ -1,0 +1,5 @@
+chef_ingredient 'chef-server' do
+  action :install
+  channel :unstable
+  version :latest
+end