+++ cookbooks/chef-ingredient/test/fixtures/cookbooks/test/recipes/unstable_chefdk.rb
@@ -1,0 +1,5 @@
+chef_ingredient 'chefdk' do
+  action :upgrade
+  channel :unstable
+  version :latest
+end