+++ cookbooks/chef-ingredient/test/fixtures/cookbooks/test/recipes/upgrade.rb
@@ -1,0 +1,6 @@
+chef_ingredient 'chef-server'
+
+chef_ingredient 'chef-server' do
+  action :upgrade
+  channel :current
+end