+++ cookbooks/chef-ingredient/test/integration/helpers/serverspec/spec_helper.rb
@@ -1,0 +1,7 @@
+require 'serverspec'
+
+if !(RUBY_PLATFORM !~ /mswin|mingw|windows/)
+  set :backend, :winrm
+else
+  set :backend, :exec
+end